Got a bunch of Compose stuff working.
Now local/cloud ML functionality is working in the Compose UI.
Not sending data up to the server yet. Also, sensor stuff not working yet.
Note that this version of Compose doesn't use the real Compose compiler,
so state/Model stuff does not work correctly, thus the manual invalidation
forcing in the Compose code.
Also note that the use of coroutines would conflict with the current
Compose compiler anyway, so we'll leave the manual invalidation approach
for now.

To run the Compose activity, click on the ActionBar menu item ("Compose"),
which launches the Compose UI as a sub-activity. To go back to the old UI,
click the Back button.